решение к дз 6 урока

+++ Lesson_2/Controllers/EmployeeController.cs
@@ -7,6 +7,8 @@ using System.Collections.Generic;
 using AutoMapper;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
+using Timesheets.Validation;
+using Timesheets.Validation.Requests;
 
 namespace Timesheets.Controllers
 {
@@ -18,11 +20,42 @@ namespace Timesheets.Controllers
         private IEmployeeRepository _employeeRepository;
         private ITaskRepository _taskRepository;
         private IMapper _mapper;
-        public EmployeeController(IEmployeeRepository employeeRepository, ITaskRepository taskRepository, IMapper mapper)
+        private IGetEmployeeByIdValidator _getEmployeeByIdValidator;
+        private ICreateEmployeeValidator _getCreateEmployeeValidator;
+        private IDeleteEmployeeValidator _deleteEmployeeValidator;
+        private IGetTaskByIdValidator _getTaskByIdValidator;
+        private ICreateTaskValidator _createTaskValidator;
+        private IDeleteTaskValidator _deleteTaskValidator;
+        private ICloseTaskValidator _closeTaskValidator;
+        private IAddEmployeeToTaskValidator _addEmployeeToTaskValidator;
+        private IRemoveEmployeeFromTaskValidator _removeEmployeeFromTaskValidator;
+
+        public EmployeeController(
+            IEmployeeRepository employeeRepository, 
+            ITaskRepository taskRepository, 
+            IMapper mapper,
+            IGetEmployeeByIdValidator getEmployeeByIdValidator,
+            ICreateEmployeeValidator getCreateEmployeeValidator,
+            IDeleteEmployeeValidator deleteEmployeeValidator,
+            IGetTaskByIdValidator getTaskByIdValidator,
+            ICreateTaskValidator createTaskValidator,
+            IDeleteTaskValidator deleteTaskValidator,
+            ICloseTaskValidator closeTaskValidator,
+            IAddEmployeeToTaskValidator addEmployeeToTaskValidator,
+            IRemoveEmployeeFromTaskValidator removeEmployeeFromTaskValidator)
         {
             _employeeRepository = employeeRepository;
             _taskRepository = taskRepository;
             _mapper = mapper;
+            _getEmployeeByIdValidator = getEmployeeByIdValidator;
+            _getCreateEmployeeValidator = getCreateEmployeeValidator;
+            _deleteEmployeeValidator = deleteEmployeeValidator;
+            _getTaskByIdValidator = getTaskByIdValidator;
+            _createTaskValidator = createTaskValidator;
+            _deleteTaskValidator = deleteTaskValidator;
+            _closeTaskValidator = closeTaskValidator;
+            _addEmployeeToTaskValidator = addEmployeeToTaskValidator;
+            _removeEmployeeFromTaskValidator = removeEmployeeFromTaskValidator;
         }
 
         [HttpGet("get/all")]
@@ -46,6 +79,13 @@ namespace Timesheets.Controllers
         public async Task<IActionResult> GetById([FromRoute] long id)
         {
             var request = new GetEmployeeByIdRequest { Id = id };
+            var validation = new OperationResult<GetEmployeeByIdRequest>(_getEmployeeByIdValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
+
             var employee = await _employeeRepository.GetById(request);
             var response = new GetEmployeeByIdResponse();
             
@@ -58,6 +98,13 @@ namespace Timesheets.Controllers
         public async Task<IActionResult> Create([FromRoute] string name)
         {
             var request = new CreateEmployeeRequest { Name = name };
+            var validation = new OperationResult<CreateEmployeeRequest>(_getCreateEmployeeValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+
+            }
             await _employeeRepository.Create(request);
             return Ok();
         }
@@ -66,6 +113,13 @@ namespace Timesheets.Controllers
         public async Task<IActionResult> Delete([FromRoute] long id)
         {
             var request = new DeleteEmployeeRequest { Id = id };
+            var validation = new OperationResult<DeleteEmployeeRequest>(_deleteEmployeeValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
+
             await _employeeRepository.Delete(request);
             return Ok();
         }
@@ -91,6 +145,13 @@ namespace Timesheets.Controllers
         public async Task<IActionResult> GetTaskById([FromRoute] long id)
         {
             var request = new GetTaskByIdRequest { Id = id };
+            var validation = new OperationResult<GetTaskByIdRequest>(_getTaskByIdValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
+
             var task = await _taskRepository.GetById(request);
             var response = new GetTaskByIdResponse();
 
@@ -103,6 +164,13 @@ namespace Timesheets.Controllers
         public async Task<IActionResult> CreateTask([FromRoute] long price)
         {
             var request = new CreateTaskRequest { PricePerHour = price };
+            var validation = new OperationResult<CreateTaskRequest>(_createTaskValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
+
             await _taskRepository.Create(request);
             return Ok();
         }
@@ -111,6 +179,13 @@ namespace Timesheets.Controllers
         public async Task<IActionResult> DeleteTask([FromRoute] long id)
         {
             var request = new DeleteTaskRequest { Id = id };
+            var validation = new OperationResult<DeleteTaskRequest>(_deleteTaskValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
+
             await _taskRepository.Delete(request);
             return Ok();
         }
@@ -119,7 +194,14 @@ namespace Timesheets.Controllers
         public async Task<IActionResult> AddEmployeeToTask([FromRoute] long employeeId, [FromRoute] long id)
         {
             var request = new AddEmployeeToTaskRequest { TaskId = id, EmployeeId = employeeId };
-           await _taskRepository.AddEmployeeToTask(request);
+            var validation = new OperationResult<AddEmployeeToTaskRequest>(_addEmployeeToTaskValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
+
+            await _taskRepository.AddEmployeeToTask(request);
             return Ok();
         }
 
@@ -127,6 +209,13 @@ namespace Timesheets.Controllers
         public async Task<IActionResult> RemoveEmployeeFromTask([FromRoute] long employeeId, [FromRoute] long id)
         {
             var request = new RemoveEmployeeFromTaskRequest { TaskId = id, EmployeeId = employeeId };
+            var validation = new OperationResult<RemoveEmployeeFromTaskRequest>(_removeEmployeeFromTaskValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
+
             await _taskRepository.RemoveEmployeeFromTask(request);
             return Ok();
         }
@@ -135,6 +224,13 @@ namespace Timesheets.Controllers
         public async Task<IActionResult> CloseTask([FromRoute] long id)
         {
             var request = new CloseTaskRequest { Id = id };
+            var validation = new OperationResult<CloseTaskRequest>(_closeTaskValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
+
             await _taskRepository.CloseTask(request);
             return Ok();
         }
